Refactor to abstract out duplication in the Data classes
